Refactor  components to use new table data and adapter structure
