+++ src/app/modules/feasibility-query/components/editor/reference-criteria-modal/reference/reference.component.ts
@@ -1,14 +1,15 @@
 import { ActiveSearchTermService } from 'src/app/service/Search/ActiveSearchTerm.service';
+import { CheckboxCellData } from 'src/app/shared/models/TableData/cells/CheckboxCellData';
 import { Component, EventEmitter, Input, OnDestroy, OnInit, Output } from '@angular/core';
 import { CriteriaSetSearchService } from 'src/app/service/Search/SearchTypes/CriteriaSet/CriteriaSetSearch.service';
 import { Display } from 'src/app/model/DataSelection/Profile/Display';
 import { filter, Observable, Subscription, tap } from 'rxjs';
-import { InterfaceTableDataRow } from 'src/app/shared/models/TableData/InterfaceTableDataRows';
 import { ReferenceCriteriaListEntry } from 'src/app/model/Search/ListEntries/ReferenceCriteriaListEntry';
 import { ReferenceCriteriaListEntryAdapter } from 'src/app/shared/models/TableData/Adapter/ReferenceCriteriaListEntryAdapter';
 import { ReferenceCriteriaResultList } from 'src/app/model/Search/ResultList/ReferenceCriteriaResultList';
 import { SelectedTableItemsService } from 'src/app/service/SearchTermListItemService.service';
-import { TableData } from 'src/app/shared/models/TableData/InterfaceTableData';
+import { TableData } from 'src/app/shared/models/TableData/TableData';
+import { TableRowData } from 'src/app/shared/models/TableData/TableRowData';
 import { TerminologyCode } from 'src/app/model/Terminology/TerminologyCode';
 
 interface selectedItem {
@@ -68,7 +69,7 @@ export class ReferenceComponent implements OnInit, OnDestroy {
       )
       .subscribe((searchTermResults: ReferenceCriteriaResultList) => {
         this.listItems = searchTermResults.getResults();
-        this.adaptedData = ReferenceCriteriaListEntryAdapter.adapt(this.listItems);
+        this.adaptedData = new ReferenceCriteriaListEntryAdapter().adapt(this.listItems);
         if (this.adaptedData.body.rows.length > 0) {
           this.searchResultsFound = true;
         } else {
@@ -103,13 +104,11 @@ export class ReferenceComponent implements OnInit, OnDestroy {
 
   private uncheckAllRows(): void {
     this.adaptedData?.body.rows.forEach((item) => {
-      if (item.isCheckboxSelected) {
-        this.uncheckRow(item);
+      const checkboxCell = item.cells.find((c): c is CheckboxCellData => c.type === 'checkbox');
+      if (checkboxCell) {
+        checkboxCell.isSelected = false;
       }
     });
-  }
-  private uncheckRow(item: InterfaceTableDataRow): void {
-    item.isCheckboxSelected = false;
   }
 
   public startElasticSearch(searchtext: string) {
@@ -142,7 +141,7 @@ export class ReferenceComponent implements OnInit, OnDestroy {
     this.selectedTableItemsService.clearSelection();
   }
 
-  public setSelectedRowItem(item: InterfaceTableDataRow) {
+  public setSelectedRowItem(item: TableRowData): void {
     const selectedIds = this.selectedTableItemsService.getSelectedIds();
     const itemId = item.originalEntry.getId();
     if (selectedIds.includes(itemId)) {
