+++ src/app/modules/shared-filter/components/shared-concept-filter/concept-filter-table/concept-filter-table.component.ts
@@ -1,14 +1,15 @@
 import { ActiveSearchTermService } from 'src/app/service/Search/ActiveSearchTerm.service';
+import { CheckboxCellData } from 'src/app/shared/models/TableData/cells/CheckboxCellData';
 import { CloneConcept } from 'src/app/model/Utilities/CriterionCloner/ValueAttributeFilter/Concept/CloneConcept';
 import { CodeableConceptListEntryAdapter } from 'src/app/shared/models/TableData/Adapter/CodeableConceptListEntryAdapter';
 import { CodeableConceptResultList } from 'src/app/model/Search/ResultList/CodeableConcepttResultList';
 import { CodeableConceptResultListEntry } from 'src/app/model/Search/ListEntries/CodeableConceptResultListEntry';
 import { CodeableConceptSearchService } from 'src/app/service/Search/SearchTypes/CodeableConcept/CodeableConceptSearch.service';
 import { Component, Input, OnDestroy, OnInit } from '@angular/core';
-import { InterfaceTableDataRow } from 'src/app/shared/models/TableData/InterfaceTableDataRows';
 import { Observable, Subscription, switchMap } from 'rxjs';
 import { SelectedConceptFilterProviderService } from '../../../service/ConceptFilter/SelectedConceptFilterProvider.service';
-import { TableData } from 'src/app/shared/models/TableData/InterfaceTableData';
+import { TableData } from 'src/app/shared/models/TableData/TableData';
+import { TableRowData } from 'src/app/shared/models/TableData/TableRowData';
 
 @Component({
   selector: 'num-concept-filter-table',
@@ -45,7 +46,7 @@ export class ConceptFilterTableComponent implements OnInit, OnDestroy {
       .getSearchResults(this.valueSetUrl)
       .pipe(
         switchMap((results) => {
-          this.adaptedData = CodeableConceptListEntryAdapter.adapt(results.getResults());
+          this.adaptedData = new CodeableConceptListEntryAdapter().adapt(results.getResults());
           return this.selectedConceptProviderService.getSelectedConcepts();
         })
       )
@@ -60,9 +61,10 @@ export class ConceptFilterTableComponent implements OnInit, OnDestroy {
     this.adaptedData?.body.rows.forEach((row) => {
       const listEntry = row.originalEntry as CodeableConceptResultListEntry;
       const concept = CloneConcept.deepCopyConcept(listEntry.getConcept());
-      row.isCheckboxSelected = this.selectedConceptProviderService.findConcept(concept)
-        ? true
-        : false;
+      const checkboxCell = row.cells.find((c): c is CheckboxCellData => c.type === 'checkbox');
+      if (checkboxCell) {
+        checkboxCell.isSelected = !!this.selectedConceptProviderService.findConcept(concept);
+      }
     });
   }
 
@@ -71,7 +73,7 @@ export class ConceptFilterTableComponent implements OnInit, OnDestroy {
     this.subscription2?.unsubscribe();
   }
 
-  public addSelectedRow(item: InterfaceTableDataRow) {
+  public addSelectedRow(item: TableRowData): void {
     const entry = item.originalEntry as CodeableConceptResultListEntry;
     const concept = CloneConcept.deepCopyConcept(entry.getConcept());
     this.selectedConceptProviderService.addConcept(concept);
